refactor(auto-pairs): consolidate pair terminology and types
