+++ src/core_editor/editor.rs
@@ -232,14 +232,14 @@ impl Editor {
             EditCommand::InsertChar(c) => self.insert_char(*c),
             EditCommand::Complete => {}
             EditCommand::InsertString(str) => self.insert_str(str),
-            EditCommand::InsertPair { left, right } => self.insert_pair(*left, *right),
+            EditCommand::InsertPair { open, close } => self.insert_pair(*open, *close),
             EditCommand::InsertNewline => self.insert_newline(),
             EditCommand::InsertNewlineAbove => self.insert_newline_above(),
             EditCommand::InsertNewlineBelow => self.insert_newline_below(),
             EditCommand::ReplaceChar(chr) => self.replace_char(*chr),
             EditCommand::ReplaceChars(n_chars, str) => self.replace_chars(*n_chars, str),
             EditCommand::Backspace => self.backspace(),
-            EditCommand::BackspacePair { left, right } => self.backspace_pair(*left, *right),
+            EditCommand::BackspacePair { open, close } => self.backspace_pair(*open, *close),
             EditCommand::Delete => self.delete(),
             EditCommand::CutChar => self.cut_char(),
             EditCommand::CutCharLeft => self.cut_char_left(),
